after
samtools mpileup -f in.fa  in.bam -g > out.bcf

and subsequent

bcftools view out.bcf | less

I noticed several lines with:

DP=1;I16=33,0,1,0,0,0,33,1089,0,0,60,3600,0,0,13,169
...
DP=21;I16=37,15,1,1,1121,69267,53,1489,1140,68400,120,7200,227,3357,27,405

Note that the first `I16' element is greater than the `DP'. This is a bug
and occurs only in the case of an `N' base. In that case in function
bcf_call_glfgen() b becomes 4, and it writes past the boundaries
of array r->qsum[], in function bcf_call_glfgen() at the line:

r->qsum[b] += q;

Since b is 4, one past the array qsum, the value of q is added to the next
member of struct bcf_callret1_t, in this case to r->anno[0].

In this patch I increased the size of qsum to prevent the write past qsum
array boundaries and enable annotation of the `N' base qualities.

In addition the memset to zero in bcf_call_glfgen() was redundant since it
was already done a few lines earlier.

Signed-off-by: Roel Kluin <roel.kluin@gmail.com>